+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 - **[UPDATE]** Add `autocompleteFrom` & `autocompleteTo` props to `SearchForm`
 - **[UPDATE]** Add new layoutMode props to `Loader` and add support for `block` mode.
 - **[BREAKING CHANGE]** Replace autocompleteProps to be a function renderAutocomplete in search form
-  [...]
+- **[FIX]** Small UI fixes on `SearchForm`
+- [...]
 
 # v27.0.0 (09/04/2020)